+++ components/retro-go/rg_gui.h
@@ -182,10 +182,10 @@ enum colors565
 };
 
 void rg_gui_init(void);
-void rg_gui_set_font_size(size_t size);
-size_t rg_gui_get_font_size(void);
-size_t rg_gui_get_font_width(void);
-size_t rg_gui_draw_text(uint16_t x, uint16_t y, uint16_t width, const char *text, uint16_t color, uint16_t color_bg);
+void rg_gui_set_font_size(int size);
+int  rg_gui_get_font_size(void);
+int  rg_gui_get_font_width(void);
+int  rg_gui_draw_text(int x, int y, int width, const char *text, uint16_t color, uint16_t color_bg);
 void rg_gui_draw_rect(int x, int y, int width, int height, int border, uint16_t color);
 void rg_gui_draw_fill_rect(int x, int y, int width, int height, uint16_t color);
 void rg_gui_draw_battery(int x, int y);
